Improve bottleneck on open()
In this commit, I changed the type of directory children to DashMap. I also wrapped throughput measurement tests with wrapper.

In the previous implementation, children field of MemFSDirNode is protected with RwLock, which leads to multiple adds not being concurrent. DashMap was a good solution for the bottleneck.

Also, I've written a macro and function for throughput measurement tests, now it is easier to write measurement tests.